+++ RELEASENOTES.rst
@@ -3,6 +3,12 @@
 Release notes
 =============
 
+Version 0.02
+-------------
+
+* Alpha releaase of Macro node
+* Only 'url' and 'params' fields are implemented
+
 Version 0.01
 -------------